changed XML's getChild to child / getChilds to childs
and some addition

+++ src/flash/text/TextFormat.java
@@ -1,14 +1,14 @@
 package flash.text;
 
 public class TextFormat {
-	public int size;
+	public double size;
 	public Object font;
 	public String align;
 	public boolean bold;
 	public int leftMargin;
 	public int indent;
 	public int leading;
-	public int color;
+	public double color;
 
 	public TextFormat(String name, double renderingFontSize) {
